security improvements and modular design
