+++ .env
@@ -1,1 +1,2 @@
 SECRET_KEY="VCC2023LOGIN"
+ADMIN_PASSWORD="BOSSMAN"
