+++ main.py
@@ -5,6 +5,9 @@ import sqlite3
 from datetime import datetime
 from difflib import SequenceMatcher
 from fastapi import FastAPI,UploadFile, Request, HTTPException, status, File, Query, Depends ,Body
+from database import lifespan, execute_db_query, get_question, get_attempts_count, decrement_question_points, reset_question_points, update_team, update_attempted_questions, create_database
+from models import Answer, Admin, Team, Score, TeamScores, TeamsInput
+from auth import create_access_token, get_current_user, verify_password
 from fastapi.middleware.cors import CORSMiddleware
 from contextlib import asynccontextmanager
 from pydantic import BaseModel
@@ -21,40 +24,20 @@ load_dotenv()
 CURRENT_DIR = os.path.dirname(__file__)
 CURRENT_DB = os.path.join(CURRENT_DIR, "initial.db")
 SECRET_KEY = os.getenv('SECRET_KEY')
+ADMIN_PASSWORD = os.getenv('ADMIN_PASSWORD')
 ALGORITHM = "HS256"
 ACCESS_TOKEN_EXPIRE_MINUTES = 60
 oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/token")
 
 
-class Answer(BaseModel):
-    id: str
-    answer: str
-
-class Admin(BaseModel):
-    admin_password: str
-    
-class Team(BaseModel):
-    team_name: str
-    password: str
-
-class Score(BaseModel):
-    q1_score: int
-    q2_score: int
-    q3_score: int
-    q4_score: int
-
-class TeamScores(BaseModel):
-    team_name: str
-    scores: Score
-
-class TeamsInput(BaseModel):
-    teams: List[TeamScores]
-
+# main.py
 @asynccontextmanager
 async def lifespan(app: FastAPI):
     with open(os.path.join(CURRENT_DIR, "initial.json"), 'r') as f:
         initial_data = json.load(f)
-    create_database(initial_data)
+    teams_json_path = os.path.join(CURRENT_DIR, "teams.json")
+    colors_json_path = os.path.join(CURRENT_DIR, "colors.json")
+    create_database(initial_data, teams_json_path, colors_json_path)
     yield
 
 
@@ -76,25 +59,6 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
-# Database function
-def execute_db_query(query, params=(), fetchone=False, db=None):
-    if db is None:
-        db=CURRENT_DB
-    try:
-        conn = sqlite3.connect(db)
-        c = conn.cursor()
-        c.execute(query, params)
-        conn.commit()
-        if fetchone:
-            return c.fetchone()
-        else:
-            return c.fetchall()
-    except Exception as e:
-        logging.error("Error occurred when executing database query", exc_info=True)
-        raise e
-    finally:
-        conn.close()
-
 
 def similar(s1, s2, threshold=0.6):
     if s1.isalpha() and s2.isalpha():
@@ -105,49 +69,6 @@ def similar(s1, s2, threshold=0.6):
     similarity_ratio = SequenceMatcher(None, s1, s2).ratio()
     #if s1 and s2 are strings and not numeric
     return similarity_ratio >= threshold
-
-def get_question(id: str):
-    result = execute_db_query("SELECT answer,current_points FROM questions WHERE id = ?", (id, ))
-    if not result:
-        raise HTTPException(status_code=404, detail="Question not found")
-    return result[0]
-
-def get_attempts_count(team_name: str,id: str):
-    count = execute_db_query("SELECT COUNT(*) FROM attempted_questions WHERE team_name = ? AND question_id = ?", (team_name, id,))
-    return count[0][0]
-
-def decrement_question_points(question_id: int):
-    execute_db_query("UPDATE questions SET current_points = current_points - 1 WHERE id = ?", (question_id,))
-
-def reset_question_points():
-    execute_db_query("UPDATE questions SET current_points = original_points")
-
-def update_team(name: str, points: int):
-    execute_db_query("UPDATE teams SET score = score + ? WHERE name = ?", (points, name))
-
-def update_attempted_questions(name: str, question_id: str, solved: bool):
-    execute_db_query(
-        f"INSERT INTO attempted_questions VALUES (?, ?, ?, ?)",
-        params=(name, question_id, datetime.now(), solved)
-    )
-
-
-def create_access_token(data: dict, expires_delta: timedelta = None):
-    to_encode = data.copy()
-    expire = datetime.now() + expires_delta if expires_delta else datetime.now() + timedelta(minutes=15)
-    to_encode.update({"exp": expire})
-    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
-    return encoded_jwt
-
-def get_current_user(token: str = Depends(oauth2_scheme)):
-    try:
-        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
-        team_name: str = payload.get("sub")
-        if team_name is None:
-            raise HTTPException(status_code=401, detail="Invalid token")
-    except JWTError:
-        raise HTTPException(status_code=401, detail="Invalid token")
-    return team_name
 
 
 @app.get("/get_comp_table")
@@ -186,7 +107,7 @@ async def get_comp_table():
 async def manual_questions(admin_password:str):
     if not admin_password:
         return {"status": "failed", "message": "Admin credentials are wrong"}
-    if admin_password != "BOSSMAN":
+    if admin_password != ADMIN_PASSWORD:
         return {"status": "failed", "message": "Admin credentials are wrong"}
     else:
     
@@ -207,7 +128,10 @@ async def manual_questions(admin_password:str):
         return teams
 
 @app.get("/questions")
-async def get_questions(team_name: str = Depends(get_current_user)):
+async def get_questions(current_user: dict = Depends(get_current_user)):
+    if current_user["role"] != "student":
+        raise HTTPException(status_code=403, detail="Forbidden")
+    team_name = current_user["team_name"]
     result = execute_db_query("SELECT * FROM teams WHERE name = ?",(team_name,))
     if not result:
         return {"status": "failed", "message": "Team credentials are wrong"}
@@ -259,7 +183,10 @@ async def get_questions(team_name: str = Depends(get_current_user)):
         return {"questions": "Error"}
 
 @app.post("/submit_mcqs_answer")
-async def submit_answer_mcqs(a: Answer, team_name: str = Depends(get_current_user)):
+async def submit_answer_mcqs(a: Answer, current_user: dict = Depends(get_current_user)):
+    if current_user["role"] != "student":
+        raise HTTPException(status_code=403, detail="Forbidden")
+    team_name = current_user["team_name"]
     result = execute_db_query("SELECT * FROM teams WHERE name = ?",(team_name,))
     if not result:
         return {"status": "failed", "message": "Team credentials are wrong"}
@@ -283,7 +210,10 @@ async def submit_answer_mcqs(a: Answer, team_name: str = Depends(get_current_use
         return {"message":"An error occurred when submitting the answer."}
 
 @app.post("/submit_sa_answer")
-async def submit_answer_sa(a: Answer, team_name: str = Depends(get_current_user)):
+async def submit_answer_sa(a: Answer, current_user: dict = Depends(get_current_user)):
+    if current_user["role"] != "student":
+        raise HTTPException(status_code=403, detail="Forbidden")
+    team_name = current_user["team_name"]
     result = execute_db_query("SELECT * FROM teams WHERE name = ?",(team_name,))
     if not result:
         return {"status": "failed", "message": "Team credentials are wrong"}
@@ -311,25 +241,33 @@ async def submit_answer_sa(a: Answer, team_name: str = Depends(get_current_user)
 @app.post("/team_login")
 async def team_login(user: Team):
     try:
-        result = execute_db_query("SELECT password FROM teams WHERE name=?",(user.team_name,))
-        if result and result[0][0]==user.password:
-            access_token_expires = timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
-            access_token = create_access_token(
-                data={"sub": user.team_name}, 
-                expires_delta=access_token_expires
-            )
-            return {"access_token": access_token, "token_type": "bearer"}
-        else:
-            return {"status": "failed", "message": "No team found with these credentials"}
-            
+        print("Trying to execute database query...")
+        result = execute_db_query("SELECT password_hash FROM teams WHERE name=?",(user.team_name,))
+        print("Database query executed.")
+        if result:
+            print("Result found.")
+            hashed_password = result[0][0]
+            if verify_password(user.password, hashed_password):
+                print("Password verified.")
+                access_token_expires = timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
+                access_token = create_access_token(
+                    data={"sub": user.team_name, "role": "student"}, 
+                    expires_delta=access_token_expires
+                )
+                print("Access token created.")
+                return {"access_token": access_token, "token_type": "bearer"}
+        print("No team found with the provided credentials.")
+        return {"status": "failed", "message": "No team found with these credentials"}
     except Exception as e:
-         return {"status": "failed", "message": "Server error"}
+        print("An exception occurred:", str(e))
+        return {"status": "failed", "message": "Server error"}
+            
 
 @app.post("/team_signup/")
 async def quick_signup(team: Team,a: Admin):
     if not a.admin_password:
         return {"status": "failed", "message": "Admin credentials are wrong"}
-    if a.admin_password != "BOSSMAN":
+    if a.admin_password != ADMIN_PASSWORD:
         return {"status": "failed", "message": "Admin credentials are wrong"}
     else:
         try:
@@ -352,7 +290,7 @@ async def quick_signup(team: Team,a: Admin):
 async def list_json_files(admin_password:str):
     if not admin_password:
         return {"status": "failed", "message": "Admin credentials are wrong"}
-    if admin_password != "BOSSMAN":
+    if admin_password != ADMIN_PASSWORD:
         return {"status": "failed", "message": "Admin credentials are wrong"}
     else:
         try:
@@ -367,7 +305,7 @@ async def set_json(filename: str,a: Admin):
     global CURRENT_DB
     if not a.admin_password:
         return {"status": "failed", "message": "Admin credentials are wrong"}
-    if a.admin_password != "BOSSMAN":
+    if a.admin_password != ADMIN_PASSWORD:
         return {"status": "failed", "message": "Admin credentials are wrong"}
     else:
         if filename:
@@ -377,11 +315,13 @@ async def set_json(filename: str,a: Admin):
             return {"status": "failed", "message": "Wrong file selected!"}
 
 @app.post("/reset_rankings/")
-async def reset_team_data(team_name: str = Query(None, description="The name of the team to reset. If not provided, all teams will be reset."),a: Admin = Body(..., embed=True)):
+async def reset_team_data(team_name: str = Query(None, description="The name of the team to reset. If not provided, all teams will be reset."), current_user: dict = Depends(get_current_user)):
+    if current_user["role"] != "admin":
+        raise HTTPException(status_code=403, detail="Forbidden")
     try:
         if not a.admin_password:
             return {"status": "failed", "message": "Admin credentials are wrong"}
-        if a.admin_password != "BOSSMAN":
+        if a.admin_password != ADMIN_PASSWORD:
             return {"status": "failed", "message": "Admin credentials are wrong"}
         else:
             # Reset stats for a specific team
@@ -425,11 +365,13 @@ async def reset_team_data(team_name: str = Query(None, description="The name of 
         return {"status": "failed", "message": "Cannot reset due to an error"}
 
 @app.post("/reset_questions_score/")
-async def reset_questions_score(a: Admin):
+async def reset_questions_score(current_user: dict = Depends(get_current_user)):
+    if current_user["role"] != "admin":
+        raise HTTPException(status_code=403, detail="Forbidden")
     try:
         if not a.admin_password:
             return {"status": "failed", "message": "Admin credentials are wrong"}
-        if a.admin_password != "BOSSMAN":
+        if a.admin_password != ADMIN_PASSWORD:
             return {"status": "failed", "message": "Admin credentials are wrong"}
         else:
             reset_question_points()
@@ -442,7 +384,7 @@ async def reset_questions_score(a: Admin):
 async def admin_login(a: Admin):
     if not a.admin_password:
         return {"status": "failed", "message": "Admin credentials are wrong"}
-    if a.admin_password != "BOSSMAN":
+    if a.admin_password != ADMIN_PASSWORD:
         return {"status": "failed", "message": "Admin credentials are wrong"}
     else:
         return {"status": "success"}
@@ -450,43 +392,42 @@ async def admin_login(a: Admin):
 
 
 @app.post("/update_manual_score/")
-async def update_manual_score(data: TeamsInput,a: Admin):
-    if not a.admin_password:
-        return {"status": "failed", "message": "Admin credentials are wrong"}
-    if a.admin_password != "BOSSMAN":
-        return {"status": "failed", "message": "Admin credentials are wrong"}
-    else:
-        try:
-            for team in data.teams:
-                scores = team.scores
-                # Check conditions for scores
-                if not 0 <= scores.q1_score <= 600:
-                    return {"status": "failed", "message": "q1_score out of range for team: " + team.team_name}
-                if not 0 <= scores.q2_score <= 600:
-                    return {"status": "failed", "message": "q2_score out of range for team: " + team.team_name}
-                if not 0 <= scores.q3_score <= 1000:
-                    return {"status": "failed", "message": "q3_score out of range for team: " + team.team_name}
-                if not -10000 <= scores.q4_score <= 10000:
-                    return {"status": "failed", "message": "q4_score out of range for team: " + team.team_name}
-                
-            for team in data.teams:
-                team_name = team.team_name
-                scores = team.scores
-                execute_db_query("""UPDATE manual_scores
-                            SET q1_score = ?, q2_score = ?, q3_score = ?, q4_score = ?
-                            WHERE team_name = ?""", (scores.q1_score, scores.q2_score, scores.q3_score, scores.q4_score, team_name,))
+async def update_manual_score(data: TeamsInput, current_user: dict = Depends(get_current_user)):
+    if current_user["role"] != "admin":
+        raise HTTPException(status_code=403, detail="Forbidden")
+    try:
+        for team in data.teams:
+            scores = team.scores
+            # Check conditions for scores
+            if not 0 <= scores.q1_score <= 600:
+                return {"status": "failed", "message": "q1_score out of range for team: " + team.team_name}
+            if not 0 <= scores.q2_score <= 600:
+                return {"status": "failed", "message": "q2_score out of range for team: " + team.team_name}
+            if not 0 <= scores.q3_score <= 1000:
+                return {"status": "failed", "message": "q3_score out of range for team: " + team.team_name}
+            if not -10000 <= scores.q4_score <= 10000:
+                return {"status": "failed", "message": "q4_score out of range for team: " + team.team_name}
             
-            return {"status": "success"}
-        except Exception as e:
-            return {"status": "failed"}
+        for team in data.teams:
+            team_name = team.team_name
+            scores = team.scores
+            execute_db_query("""UPDATE manual_scores
+                        SET q1_score = ?, q2_score = ?, q3_score = ?, q4_score = ?
+                        WHERE team_name = ?""", (scores.q1_score, scores.q2_score, scores.q3_score, scores.q4_score, team_name,))
+        
+        return {"status": "success"}
+    except Exception as e:
+        return {"status": "failed"}
 
 
 @app.post("/upload/{admin_password}")
-async def upload_database(admin_password:str,file: UploadFile = File(None)):
+async def upload_database(admin_password: str, file: UploadFile = File(None), current_user: dict = Depends(get_current_user)):
+    if current_user["role"] != "admin":
+        raise HTTPException(status_code=403, detail="Forbidden")
     global CURRENT_DB
     if not admin_password:
         return {"status": "failed", "message": "Admin credentials are wrong"}
-    if admin_password != "BOSSMAN":
+    if admin_password != ADMIN_PASSWORD:
         return {"status": "failed", "message": "Admin credentials are wrong"}
     else:
         if file:
@@ -516,125 +457,3 @@ async def upload_database(admin_password:str,file: UploadFile = File(None)):
                 raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail=str(e))
         else:
             return {"status": "failed", "message": "File Not uploaded"}
-
-
-def create_database(data):
-    db_file_path = os.path.join(CURRENT_DIR, f"{CURRENT_DB}")
-
-    # Delete the database file if it already exists
-    if os.path.exists(db_file_path):
-        os.remove(db_file_path)
-    try:
-        conn = sqlite3.connect(db_file_path)
-        cursor = conn.cursor()
-        
-        # Define tables
-        teams_table = """
-        CREATE TABLE "teams" (
-            "name"	TEXT NOT NULL UNIQUE,
-            "password"	TEXT NOT NULL,
-            "score"	INTEGER DEFAULT 0,
-            "color"	TEXT,
-            PRIMARY KEY("name")
-        );
-        """
-        questions_table = """
-        CREATE TABLE "questions" (
-            "id"	INTEGER,
-            "title"     TEXT,
-            "content"	TEXT NOT NULL,
-            "answer"	TEXT NOT NULL,
-            "original_points"	INTEGER NOT NULL,
-            "current_points"	INTEGER,
-            "type"	TEXT,
-            "question_group"	INTEGER,
-            "option_a"	TEXT,
-            "option_b"	TEXT,
-            "option_c"	TEXT,
-            "option_d"	TEXT,
-            "option_e"	TEXT,
-            "option_f"	TEXT,
-            "option_g"	TEXT,
-            "option_h"	TEXT,
-            "option_i"	TEXT,
-            "option_j"	TEXT,
-            "image_link"	TEXT,
-            "content_link"	TEXT,
-            PRIMARY KEY("id")
-        );
-        """
-        attempted_questions_table = """
-        CREATE TABLE "attempted_questions" (
-            "team_name"	text,
-            "question_id"	INTEGER,
-            "timestamp"	datetime,
-            "solved"	boolean NOT NULL,
-            FOREIGN KEY("team_name") REFERENCES "teams"("name"),
-            FOREIGN KEY("question_id") REFERENCES "questions"("id")
-            PRIMARY KEY("team_name", "timestamp")
-        );
-        """
-        manual_question_table = """
-        CREATE TABLE "manual_scores" (
-            "team_name"	TEXT UNIQUE,
-            "q1_score"	INTEGER,
-            "q2_score"	INTEGER,
-            "q3_score"	INTEGER,
-            "q4_score"  INTEGER,
-            FOREIGN KEY("team_name") REFERENCES "teams"("name")
-        );
-        """
-
-        cursor.execute(teams_table)
-        cursor.execute(questions_table)
-        cursor.execute(attempted_questions_table)
-        cursor.execute(manual_question_table)
-
-        for question in data['questions']:
-            cursor.execute(
-                "INSERT INTO questions (content, title, answer, original_points, current_points, type, question_group, option_a, option_b, option_c, option_d, option_e, option_f, option_g, option_h, option_i, option_j, image_link, content_link) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                (
-                    question['content'],
-                    question['title'],
-                    question['answer'],
-                    question['original_points'],
-                    question['original_points'],
-                    question['type'],
-                    question['question_group'],
-                    question.get('option_a', None),
-                    question.get('option_b', None),
-                    question.get('option_c', None),
-                    question.get('option_d', None),
-                    question.get('option_e', None),
-                    question.get('option_f', None),
-                    question.get('option_g', None),
-                    question.get('option_h', None),
-                    question.get('option_i', None),
-                    question.get('option_j', None),
-                    question.get('image_link', None),
-                    question.get('content_link', None)
-                )
-            )
-        teams_json_path = os.path.join(CURRENT_DIR, 'teams.json')
-        colors_json_path = os.path.join(CURRENT_DIR, 'colors.json')
-
-        with open(teams_json_path, 'r') as file:
-            data = json.load(file)
-            teams_list = data['teams']
-
-        with open(colors_json_path, 'r') as file:
-            data = json.load(file)
-            color_list = data["colors"]*5
-        for team in teams_list:
-            cursor.execute("INSERT INTO teams (name, password, score, color) VALUES (?,?,?,?)",(team["name"], team["password"], 0, color_list.pop(0)))
-
-        cursor.execute("""INSERT INTO manual_scores (team_name, q1_score, q2_score, q3_score, q4_score)
-        SELECT name, 0, 0, 0, 0 FROM teams
-        WHERE name NOT IN (SELECT team_name FROM manual_scores);""")
-
-        conn.commit()
-        conn.close()
-    except Exception as e:
-        logging.error("An error occurred when creating the database", exc_info=True)
-        raise e
-
